xiph demux: _correctly_ remove warning
Thanks, j-b

+++ modules/demux/xiph.h
@@ -48,8 +48,10 @@ static inline int xiph_SplitHeaders(unsigned packet_size[], void *packet[], unsi
         }
         size += packet_size[i];
     }
-    if (end - current < (int)size)
+
+    if ((uint8_t)(end - current) < size)
         return VLC_EGENERIC;
+
     packet_size[count - 1] = end - current - size;
 
     /* Copy the payloads */
